# docs/api-spec.yaml
openapi: 3.1.0
info:
  title: AI Document Assistant API
  version: 0.1.0
  description: Versioned REST API for auth, ingestion, retrieval, audit, guardrails, and runtime stats.
servers:
  - url: http://localhost:8000
paths:
...
  /api/v1/documents/upload:
    post:
      summary: Upload one or more documents
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [files]
              properties:
                files:
                  type: array
                  items:
                    type: string
                    format: binary
      responses:
        "200":
          description: Batch upload result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DocumentBatchUploadResponse"
        "400":
          description: No files provided or file was empty
        "413":
          description: Uploaded file exceeded the configured size limit
        "415":
          description: Uploaded file type is unsupported
...
  /api/v1/documents/{document_id}:
    get:
      summary: Get document detail
      parameters:
        - in: path
          name: document_id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Document record
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DocumentRecord"
    delete:
      summary: Delete a document
      parameters:
        - in: path
          name: document_id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Deleted
  /api/v1/query:
    post:
      summary: Ask a grounded question
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/QueryRequest"
      responses:
        "200":
          description: Grounded answer
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QueryResponse"
  /api/v1/query/stream:
    post:
      summary: Stream a grounded question response
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/QueryRequest"
      responses:
        "200":
          description: NDJSON stream of delta and final response events
          content:
            application/x-ndjson:
              schema:
                type: string
...
components:
  schemas:
...
    DocumentStatus:
      type: string
      enum: [queued, processing, completed, warning, failed]
    DocumentRecord:
      type: object
      required:
        [id, filename, original_filename, content_type, size_bytes, fingerprint, status, index_status, page_count, chunk_count, warnings, created_at, updated_at]
      properties:
        id:
          type: string
        filename:
          type: string
        original_filename:
          type: string
        content_type:
          type: string
        size_bytes:
          type: integer
        fingerprint:
          type: string
        status:
          $ref: "#/components/schemas/DocumentStatus"
        index_status:
          type: string
        source_path:
          type: [string, "null"]
        duplicate_of:
          type: [string, "null"]
        page_count:
          type: integer
        chunk_count:
          type: integer
        warnings:
          type: array
          items:
            type: string
        error_message:
          type: [string, "null"]
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        indexed_at:
          type: [string, "null"]
          format: date-time
    DocumentUploadResponse:
      type: object
      required: [filename, created, duplicate, warning, chunks_created]
      properties:
        filename:
          type: string
        document:
          oneOf:
            - $ref: "#/components/schemas/DocumentRecord"
            - type: "null"
        created:
          type: boolean
        duplicate:
          type: boolean
        warning:
          type: boolean
        chunks_created:
          type: integer
        message:
          type: [string, "null"]
        error:
          type: [string, "null"]
    DocumentBatchUploadResponse:
      type: object
      required: [results, processed_count, created_count, warning_count, duplicate_count, failed_count]
      properties:
        results:
          type: array
          items:
            $ref: "#/components/schemas/DocumentUploadResponse"
        processed_count:
          type: integer
        created_count:
          type: integer
        warning_count:
          type: integer
        duplicate_count:
          type: integer
        failed_count:
          type: integer
        message:
          type: [string, "null"]
...
    QueryRequest:
      type: object
      required: [question, top_k]
      properties:
        question:
          type: string
        top_k:
          type: integer
          minimum: 1
          maximum: 10
    Citation:
      type: object
      required: [source, relevance_score]
      properties:
        source:
          type: string
        page:
          type: [integer, "null"]
        chunk_id:
          type: [string, "null"]
        relevance_score:
          type: number
        excerpt:
          type: [string, "null"]
    QueryResponse:
      type: object
      required: [answer, citations, confidence, finish_reason]
      properties:
        answer:
          type: string
        citations:
          type: array
          items:
            $ref: "#/components/schemas/Citation"
        confidence:
          type: number
        finish_reason:
          type: string
        disclaimer:
          type: [string, "null"]
...